+++ tests/unit_tests/test_history_manager.py
@@ -15,7 +15,7 @@
 import torch
 from omegaconf import OmegaConf
 
-from rlinf.workers.env.history_manager import HistoryManager
+from rlinf.utils.env_helpers import HistoryManager
 
 
 def _history_cfg():
