+++ tests/unit_tests/test_overlap_env_bootstrap.py
@@ -32,8 +32,8 @@ if "rlinf.envs.wrappers" not in sys.modules:
     sys.modules["rlinf.envs.wrappers"] = MagicMock()
 
 from rlinf.scheduler.hardware.accelerators.accelerator import AcceleratorType
+from rlinf.utils.env_helpers import SmoothInterveneController  # noqa: E402
 from rlinf.workers.env.env_worker import EnvWorker  # noqa: E402
-from rlinf.workers.env.smooth_intervene import SmoothInterveneController  # noqa: E402
 
 
 class TestOverlapEnvBootstrap(unittest.TestCase):
